dropdown dialog renamed as popup to better fit functionality
