+++ packages/Framework/src/exports/ui.ts
@@ -1,3 +1,5 @@
 export * from '../ui/layout';
 export * from '../ui/options/baseOptions';
 export * from '../ui/options/checkOptions';
+export * from '../ui/dialog';
+export * from '../ui/popup';
