+++ packages/Framework/src/ui/dialog.tsx
@@ -1,7 +1,6 @@
 import React, { useRef, memo, ReactNode } from 'react';
 import { View, StyleProp, ViewStyle } from 'react-native';
 import { Text, Modal, Button, Card, TouchableRipple, useTheme } from 'react-native-paper';
-import { Menu, MenuOptions, MenuTrigger } from 'react-native-popup-menu';
 import * as Const from '../const';
 
 type DialogProps = {
@@ -91,55 +90,3 @@ export const Dialog: React.FC<DialogProps> = memo(({
     </Modal>
   );
 });
-
-type DropdownDialogProps = {
-  triggerComp: ReactNode;
-  disabled?: boolean;
-  style?: StyleProp<ViewStyle>;
-  children: ReactNode;
-};
-
-/******************************************************************************************************************
- * Dropdown dialog component
- *
- * A general-purpose dropdown dialog component powered by react-native-popup-menu.
- * Provides:
- * - Customizable trigger component
- * - Disabled state support
- * - Flexible dropdown content area
- *
- * ⚠ NOTE: `triggerComp` must not contain its own onPress callback, as it may override option callbacks.
- * 
- * @param triggerComp - ReactNode element that triggers the dropdown when pressed
- * @param disabled? - if true, disables the trigger
- * @param style? - additional style for the container
- * @param children - content of the dropdown menu
- * 
- * @returns JSX.Element
- ******************************************************************************************************************/
-export const DropdownDialog: React.FC<DropdownDialogProps> = memo(
-  ({ triggerComp, disabled = false, style = {}, children }) => {
-    const theme = useTheme();
-    const menuRef = useRef<Menu | null>(null);
-
-    return (
-      <Menu ref={menuRef} style={style}>
-        <MenuTrigger
-          disabled={disabled}
-          customStyles={{
-            TriggerTouchableComponent: TouchableRipple,
-          }}
-        >
-          {triggerComp}
-        </MenuTrigger>
-        <MenuOptions
-          customStyles={{
-            optionsWrapper: { backgroundColor: theme.colors.surfaceVariant },
-          }}
-        >
-          {children}
-        </MenuOptions>
-      </Menu>
-    );
-  }
-);
